+++ 实例/2.1串口/code/machine.h
@@ -5,8 +5,8 @@
 #include "intrins.h"
 
 
-typedef unsigned char uchar;
-typedef unsigned int uint;
+typedef unsigned char uint8_t;
+typedef unsigned int uint16_t;
 
 
 void batch(uchar i);
